requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
gevent==21.1.2
greenlet==1.0.0
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multi-key-dict==2.0.3
pbr==5.5.1
polling==0.3.1
pycparser==2.20
pymongo==3.11.3
python-jenkins==1.7.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
